+++ CodexLib/Properties/AssemblyInfo.cs
@@ -4,8 +4,9 @@ using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("CodexLib")]
 [assembly: AssemblyProduct("CodexLib")]
+[assembly: AssemblyDescription("A shared library")]
 
-[assembly: AssemblyVersion("1.1.1.29")]
-[assembly: AssemblyFileVersion("1.1.1.29")]
+[assembly: AssemblyVersion("1.2.0.1")]
+[assembly: AssemblyFileVersion("1.2.0.1")]
 [assembly: ComVisible(false)]
 [assembly: Guid("aaa3ac08-6a5c-495f-b997-3a3b3bc4fc84")]