+++ DarkCodex/Properties/AssemblyInfo.cs
@@ -2,35 +2,11 @@ using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-// Allgemeine Informationen über eine Assembly werden über die folgenden
-// Attribute gesteuert. Ändern Sie diese Attributwerte, um die Informationen zu ändern,
-// die einer Assembly zugeordnet sind.
 [assembly: AssemblyTitle("DarkCodex")]
-// [assembly: AssemblyDescription("")]
-// [assembly: AssemblyConfiguration("")]
-// [assembly: AssemblyCompany("Truinto")]
 [assembly: AssemblyProduct("DarkCodex")]
-// [assembly: AssemblyCopyright("Copyright © Truinto 2022")]
-// [assembly: AssemblyTrademark("")]
-// [assembly: AssemblyCulture("")]
+[assembly: AssemblyDescription("Mod for Pathfinder: Wrath of the Righteous")]
 
-// Durch Festlegen von ComVisible auf FALSE werden die Typen in dieser Assembly
-// für COM-Komponenten unsichtbar.  Wenn Sie auf einen Typ in dieser Assembly von
-// COM aus zugreifen müssen, sollten Sie das ComVisible-Attribut für diesen Typ auf "True" festlegen.
-[assembly: ComVisible(false)]
-
-// Die folgende GUID bestimmt die ID der Typbibliothek, wenn dieses Projekt für COM verfügbar gemacht wird
-[assembly: Guid("1b87b8ea-001b-4fe4-9993-77c33bffba87")]
-
-// Versionsinformationen für eine Assembly bestehen aus den folgenden vier Werten:
-//
-//      Hauptversion
-//      Nebenversion
-//      Buildnummer
-//      Revision
-//
-// Sie können alle Werte angeben oder Standardwerte für die Build- und Revisionsnummern verwenden,
-// indem Sie "*" wie unten gezeigt eingeben:
-// [assembly: AssemblyVersion("1.4.1")]
 [assembly: AssemblyVersion("1.4.1")]
 [assembly: AssemblyFileVersion("1.4.1")]
+[assembly: ComVisible(false)]
+[assembly: Guid("1b87b8ea-001b-4fe4-9993-77c33bffba87")]