+++ DarkCodex/Unlock.cs
@@ -55,7 +55,7 @@ namespace DarkCodex
 
         }
 
-        [PatchInfo(Severity.Create | Severity.WIP, "Unlock Kineticist", "adds infusion, wild talent, and element focus up to level 40", true, Requirement: typeof(Patch_UnlockClassLevels))]
+        [PatchInfo(Severity.Create, "Unlock Kineticist", "adds infusion, wild talent, and element focus up to level 40", true, Requirement: typeof(Patch_UnlockClassLevels))]
         public static void UnlockKineticist()
         {
             var t = KineticistTree.Instance;